+++ src/customers/mappers/map-customer-details.ts
